Pattern matching via recursive Pattern enum, u32 egraph IDs
- Replace self.patterns HashMap with compiled Pattern tree (Pattern::Var/Ground/App)
- Rewrite match_term to recursively match patterns against function_maps
- Egraph IDs now use u32 (distinct from u64 solver UIDs) for type safety
- Fix arithmetic coefficient ID space: Coefficient::Term uses egraph IDs consistently
- Fix lialp.rs/z3lp.rs proof_forest indexing to use egraph IDs
- Add early-return in insert_predecessor for pattern-allocated constants

+++ src/egraphs/congruence_closure.rs
@@ -8,10 +8,10 @@ use crate::egraphs::proofforest::ProofForestEdge;
 
 pub fn add_parent(
     proof_parent: ProofForestEdge,
-    parent: u64,
-    new_child: u64,
+    parent: u32,
+    new_child: u32,
     level: usize,
-    hash: u64,
+    hash: u32,
 ) -> ProofForestEdge {
     match proof_parent {
         ProofForestEdge::Root {
@@ -57,7 +57,7 @@ pub fn add_parent(
     }
 }
 
-pub fn get_parent(proof_parent: &ProofForestEdge) -> u64 {
+pub fn get_parent(proof_parent: &ProofForestEdge) -> u32 {
     debug_println!(6, 0, "We are getting the parent of {:?}", proof_parent);
     match proof_parent {
         ProofForestEdge::Root { .. } => {
@@ -74,7 +74,7 @@ pub fn get_parent(proof_parent: &ProofForestEdge) -> u64 {
     }
 }
 
-pub fn get_child(proof_parent: &ProofForestEdge) -> u64 {
+pub fn get_child(proof_parent: &ProofForestEdge) -> u32 {
     match proof_parent {
         ProofForestEdge::Root {
             child,
